syncing up data on deletion and creation of objects

+++ PPI-projektas/Services/NoteService.cs
@@ -1,4 +1,5 @@
-﻿using PPI_projektas.Services.Interfaces;
+﻿using PPI_projektas.Exceptions;
+using PPI_projektas.Services.Interfaces;
 using PPI_projektas.objects.Factories;
 using PPI_projektas.Services.Response;
 using PPI_projektas.Utils;
@@ -35,8 +36,10 @@ public class NoteService : INoteService
     public Guid CreateNote(Guid groupId, Guid authorId)
     {
         var group = DataHandler.FindObjectById(groupId, DataHandler.Instance.AllGroups);
+        var author = DataHandler.FindObjectById(authorId, DataHandler.Instance.AllUsers);
         var note = _noteFactory.Create(authorId);
         DataHandler.Create(note);
+        author.AddCreatedNote(note);
         group.AddNote(note);
 
         return note.Id;
@@ -55,9 +58,15 @@ public class NoteService : INoteService
     public void DeleteNote(Guid noteId, Guid userId)
     {
         var note = DataHandler.FindObjectById(noteId, DataHandler.Instance.AllNotes);
+        var user = DataHandler.FindObjectById(userId, DataHandler.Instance.AllUsers);
+        
+        var group = DataHandler.Instance.AllGroups.FirstOrDefault(group => group.Notes.Contains(note));
+        if (group == null) throw new ObjectDoesNotExistException();
         
         if (note.AuthorId != userId) throw new UnauthorizedAccessException();
         
+        group.RemoveNote(note);
+        user.RemoveCreatedNote(note);
         DataHandler.Delete(note);
     }
 }
